Fix performance regressions introduced by accessibility flags
Our performance regression was introduced by master changing
accessibilityElementsHidden from a string value to NO, changing it from
what must have been a truthy value in objc (perhaps false is represented
by the nullpointer, and true by anything not nullpointer). This makes
iOS look at all of the maps children so they can be accessed by
voiceover and friends. This turns out to be incredibly laggy for many
pins, so we revert to not having our pins be accessible.

https://developer.apple.com/documentation/objectivec/nsobject/1615141-isaccessibilityelement?language=objc
https://developer.apple.com/documentation/objectivec/nsobject/1615080-accessibilityelementshidden?language=objc

Additionally, we switch to using the newly exposed icon api instead of
views as map children. It is more performant as no View-jugglery and
measuring is required anymore. We replace the image loading in
setIconSrc with the one hardcoded for native resources as well. Using
native resources here requires us to update some of the code in
MapMarker to not try to resolve our resources before passing them to
native-land.